[Silver II] Title: 최소 힙, Time: 308 ms, Memory: 27092 KB -BaekjoonHub

+++ Java/백준/Silver/1927. 최소 힙/최소 힙.java
@@ -6,11 +6,11 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringBuilder sb = new StringBuilder();
         int n = Integer.parseInt(br.readLine());
-        PriorityQueue<Long> q = new PriorityQueue<>();
+        PriorityQueue<Integer> q = new PriorityQueue<>();
 
         int zero = 0;
         for (int i = 0; i < n; i++) {
-            long a = Long.parseLong(br.readLine());
+            int a = Integer.parseInt(br.readLine());
             if (!(a==0)) q.add(a);
             else if (a==0 && !q.isEmpty()) sb.append(q.poll() + "\n");
             else if (a==0 && q.isEmpty()) sb.append(0 + "\n");
